Move ec stuff to ec_key_pair.go
